fix: remove unneeded casting

+++ test/test_torchdynamo_bert.py
@@ -81,12 +81,9 @@ def test_benchmark_implementations(benchmark, model_reference_fp32, shape: (int,
 
     torchdynamo.reset()
 
-    assert torch.allclose(
-        input=value["last_hidden_state"].float(), other=expected["last_hidden_state"].float(), rtol=1e-1, atol=1e-1
-    )
-    assert torch.allclose(
-        input=value["pooler_output"].float(), other=expected["pooler_output"].float(), rtol=1e-1, atol=1e-1
-    )
+    assert torch.allclose(input=value["last_hidden_state"].float(), other=expected["last_hidden_state"], rtol=1e-1,
+                          atol=1e-1)
+    assert torch.allclose(input=value["pooler_output"].float(), other=expected["pooler_output"], rtol=1e-1, atol=1e-1)
 
 
 def test_support_shape_change(model_reference_fp32):
